+++ app/seeks/sesame_mall/order_seek.rb
@@ -75,18 +75,14 @@ class SesameMall::OrderSeek
   class << self
     def whole_sync
       seek = self.new
-      seek.batch_size = 2
 
       seek.do_whole_sync(relation: SesameMall::Source::Order, key: :shop_id)
     end
 
-    def partial_sync(duration: 3.days)
+    def partial_sync(duration: 30.minutes)
       seek = self.new
+      _relation = source_records_from_seek_record(klass: SesameMall::Source::Order, duration: duration)
 
-      _time = Time.now
-      _relation = SesameMall::Source::Order.where(
-        create_time: duration.ago(_time).._time
-      )
       seek.do_partial_sync(relation: _relation)
     end
   end