+++ app/seeks/sesame_mall/product_seek.rb
@@ -58,11 +58,7 @@ class SesameMall::ProductSeek
 
     def partial_sync(duration: 30.minutes)
       seek = self.new
-
-      _time = Time.now
-      _relation = SesameMall::Source::Product.where(
-        UPDATE_DATE: duration.ago(_time).._time
-      )
+      _relation = source_records_from_seek_record(klass: SesameMall::Source::Product, duration: duration)
 
       seek.do_partial_sync(relation: _relation)
     end