+++ config/deploy/production.rb
@@ -13,7 +13,7 @@ set :sidekiq_processes, 2
 # server "example.com", user: "deploy", roles: %w{app db web}, my_property: :my_value
 # server "example.com", user: "deploy", roles: %w{app web}, other_property: :other_value
 # server "db.example.com", user: "deploy", roles: %w{db}
-server "106.14.159.184:40022", user: "ishanggang_dev", roles: %w{app db web}, my_property: :my_value
+server "106.14.159.184:40022", user: "ishanggang_dev", roles: %w{app db web whenever}, my_property: :my_value
 
 
 # role-based syntax